+++ src/oci/mysql/models/db_system_source_from_none.py
@@ -23,7 +23,7 @@ class DbSystemSourceFromNone(DbSystemSource):
 
         :param source_type:
             The value to assign to the source_type property of this DbSystemSourceFromNone.
-            Allowed values for this property are: "NONE", "BACKUP", "PITR", "IMPORTURL"
+            Allowed values for this property are: "NONE", "BACKUP", "PITR", "IMPORTURL", "DBSYSTEM"
         :type source_type: str
 
         """